refactor: move get_generator to its own file
